requirements.txt
# 核心依赖
ta-lib-bin
akshare>=1.15.59
pandas>=2.2.0
numpy>=1.26.0

# 数据处理
openpyxl>=3.1.0  # Excel导出
requests>=2.31.0

# Web界面
Flask>=3.0.0
gunicorn

# 可选依赖（图表）
matplotlib>=3.8.0
mplfinance>=0.12.10
reportlab>=4.0.0
